requirements.txt
django~=2.2.0
djangorestframework~=3.10.0
